requirements.txt
aiohttp==1.0.5
aspy.yaml==1.1.1
astroid==2.0.4
async-timeout==3.0.1
asyncio==3.4.3
atomicwrites==1.2.1
attrs==18.2.0
beautifulsoup4==4.6.3
cached-property==1.5.1
certifi==2018.8.24
cfgv==1.1.0
chardet==3.0.4
coverage==4.5.1
discord.py==0.16.12
docutils==0.14
facebook-sdk
fbchat==1.4.1
flake8==3.5.0
future==0.16.0
gitdb2==2.0.4
GitPython==2.1.11
identify==1.1.7
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
more-itertools==4.3.0
multidict==4.4.2
mypy==0.630
mypy-extensions==0.4.1
nodeenv==1.3.2
oauthlib==2.1.0
pluggy==0.7.1
praw==6.0.0
prawcore==1.0.0
pre-commit==1.11.1
py==1.6.0
pycodestyle==2.3.1
pyflakes==1.6.0
pygame==1.9.4.dev0
Pygments==2.2.0
pylint==2.1.0
pytest==3.8.1
PyTumblr==0.0.8
PyYAML==3.13
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
smmap2==2.0.4
toml==0.9.6
tox==3.4.0
tweepy==3.6.0
typed-ast==1.1.0
update-checker==0.16
urllib3==1.23
virtualenv==16.0.0
websockets==3.4
wrapt==1.10.11
